refactor: renamed 'InvestTabPanel' to 'TabPanel'.
